Update math/prob documentation
Remove DEC accuracy references, we don't support DEC.

Rename function references in comments to either the name of the
function, or a generic one if it can easily be inferred from context.

Don't use expx2 function by default. Want to confirm how required this
is and remove if it isn't considered too necessary for the accuracy.